added a new feature

+++ slaughterhouse/dill.sh
@@ -1,18 +1,33 @@
 #!usr/bin/env bash
+ 
+
+
 
 #list packages that can be upgraded:
 sudo apt list --upgradable
+read -p "Press Y to upgrade and N to abort $foo?" output
+if [[ $output = y ]] ; then 
+      #update package lists:
+      sudo apt update
 
-#update package lists:
-sudo apt update
+      #upgrade all installed packages:
+      sudo apt upgrade -y
 
-#upgrade all installed packages:
-sudo apt upgrade -y
+      #remove unnecessary packages:
+      sudo apt autoremove -y
 
-#remove unnecessary packages:
-sudo apt autoremove -y
+      #clean up package cache:
+      sudo apt clean
 
-#clean up package cache:
-sudo apt clean
+      echo "============================system update and cleanup                        complete=========================="
 
-echo "============================system update and cleanup complete===================================="
+else
+      echo " "
+      echo "GOODBYE"
+fi
+      
+
+
+
+ 
+
